[Feature]: add condition to ir.sequence, let it be partitioned.
With this feature, multiple sequences may be defined per type. A weight
parameter will primarily order them. Then, the one whose condition is
true will be used.
This would let us write modules (calls to get() must be upgraded), where
the sequence numbering will depend on the data. Note that this is not
straghtforward, as the context passed must be carefully constructed. One
problem is that not all object data (eg category) may exist when the
sequence is called for a /new/ object (as a call for 'default').

The current code should be backwards compatible with old calls to get()
without a context.

bzr revid: p_christ@hol.gr-20090909203517-qahww8hika9t3zl8