doc ftp: enable ftp even before the module is fully installed.
This is marginally correct, but the document_ftp module would cause
its own tests to fail (because it's not installed yet) and thus won't
allow access to the db. Then, a 'tests.fatal' setting means that the
module will never install successfully!

Ideally, the tests should be performed *after* the db is inited and
module is fully loaded. But the yaml mechanism performs them too early
for us.

bzr revid: p_christ@hol.gr-20100827073917-g6jwwskgpnnwl855