Doc WebDAV: reuse address from the socket, not the bind server.
This should solve the following: When we reply to PROPFIND requests etc,
we provide the absolute url of ourselves like:
<D:href>http://our.server:8069/webdav/mydb/Documents</D:href>
The problem is that if we bind to 0.0.0.0 and the interface, which the
client connects through, doesn't resolve from "our.server", then we are
practically redirecting the client to the wrong address.
This is expected to happen at openerp servers w/o full qualified names and
reverse resolution of their interfaces.

Requires server patch (at websrv_lib) to function, fallbacks to old code
on trouble.

bzr revid: p_christ@hol.gr-20100926163210-v8scuvhg991jbbzc