Doc webdav, caldav: fix support for the calendar-multiget report
This report is a request to fetch multiple calendar entries. The
request contains a range of URIs to fetch, which must be processed
at the dav_fs.Report level.
The report is being called with Depth: 0 , for which the library could
not perform an iterator. Hack over it and explictly set Depth:1 in our
case.

bzr revid: p_christ@hol.gr-20101012104103-eu156146jy4a75af