document: Hide resource attachments/folders from static dirs.
If some attachment or folder references a res_model/res_id, it should
NOT be visible in the root directory (where it's parent_id points to).
This solves functional problems for the root folder, as well as the
privacy of those attachments.

bzr revid: p_christ@hol.gr-20101027102323-x4h20skrfjger9z7