setup: The ssl certs should *not* be distributed.
The meaning of any SSL certificate (even self-signed) is that it uniquely
identifies the server. So, if we have a generic cert distributed with our
packaging, we break that. We could not even generate a cert at the "build"
stage of our server, because that would be included in the packages.

If anybody needs to run OpenERP with SSL, they will need to generate the
certificate at the target server, possibly using ssl-cert.cfg as a sample.

Also, the "ssl" directory under bin/ would confuse some pythonic code that
had tried to "import ssl" (eg. urllib.py).

bzr revid: p_christ@hol.gr-20101123135844-nr8k78qrmlyn19xb