l10n_gr: refactor VAT templates and codes.
So far, there is only the most common 23% VAT there, and the 19%, 21%
ones used in 2010. The rest of the VAT codes for less-taxed goods,
in-EU moves and Aegean Islands need to be filled.

Since these tax-related xml files are likely to grow, we split them in
VAT-related, root data and other taxes (to do) files.

bzr revid: p_christ@hol.gr-20110113094636-5eovg7t8iipn1naj