[FIX] res_partner: extend default security rule to show partners from ancestors companies too
This rules is not suited for most OpenERP objects, but it works
well for partners and products as these records are usually
shared between companies, and all company-specific data (such as
acounting data is set as per-company properties). A partner or
product can thus be shared cleanly between companies.
This was already possible before by just removing the company
assigned to the partner, making it globally visible.

bzr revid: odo@openerp.com-20110303155918-zjdgwnorsiex8y8u