[ADD] text, selection and boolean fields; improve logic to detect whether a field has a value
If the value returned by a field is the literal 'null' or an empty
string, then the field is empty (empty input, no valued option
selected, etc...).

On the other hand, because openerp.base.search.Field.get_value returns
a *javascript* value it can be the literals '0' or 'false'. Those are
falsy in a boolean context (for good reasons, for 'false') but they
should still count as the field having a value.

I'm guessing this part is going to reach "spin off in its own
per-field method" complexity pretty soon.

bzr revid: xmo@openerp.com-20110324184713-sf7skv34s9824ge6