[FIX] ir.actions: fixed and documented a bit the function field computing views for actions
The previous implementation had numerous flaws and edge cases where it would
not have any useful result. The new implementation is backwards compatible
with the valid combinations of view_id, view_modes and view_id.
A unique constraint was also added to make sure that it is not possible to
specify multiple views of the same type for the same action. This was never
used, but should be prevented completely so that clients can rely on this
fact. Also a little bit of documentation was added to explain what the field
is doing.

bzr revid: odo@openerp.com-20110401124852-r0i6cclc80acoegj