[FIX] removal of unique() callbacks from callback chains
Array.prototype.pop() does *not* take a parameter, it always removes the last item of the array.

To remove items in the middle of an array, Array.prototype.splice(index, 1) has to be used.

This fixes an issue where adding a unique non-last callback (either added as non-last, or with other stuff added after it) would result in nonsensical behavior

bzr revid: xmo@openerp.com-20110628140907-5sch98gmwcnpvk3t