[FIX] mail.message: when auto-deleting a message, only delete attachments it owns
There's a peculiar way to manage attachments on
mail messages, as there is an explicit many2many
on top of regular attachments like for any
record, so it is possible to link a mail
to attachments that are owned by other records.
Such attachments must not be deleted along with
the message when the auto-delete flag is set.

bzr revid: odo@openerp.com-20120316133026-4wbrwx64nqkeyaou