[FIX] decimal_precision: properly invalidate cache when changing precisions
This was working previously because the cache was disabled,
due to bug #988743. Now that the cache is working again,
it needs to be properly invalidated whenever a decimal.precision
record is created or deleted, otherwise all computations
will use an incorrect precision.

bzr revid: odo@openerp.com-20120523085234-2q8z7f1naydnjgsz