[FIX] orm: remove unused `module` parameter when creating constraints
This argument was not always present in the context
(e.g when _auto_init is invoked for newly created
custom fields via ir.model.field.create()), and was
thus preventing the creation of custom fields.
It was currently unused, and may be reintroduced
soon after the cleanup of model metadata storage.

bzr revid: odo@openerp.com-20120606141908-bdqxsarfb61yfw8m