[FIX] product.uom: use proper ratio for uoms bigger than the reference
Bigger UoMs should logically be specified using the
`factor_inv` column, similarly to what is done in
the form view when you select the Bigger type.
This also fixes the rounding issue of Dozen, which
was erroneously specified as .083, which is far
from 1/12.

bzr revid: odo@openerp.com-20121003115732-ws73com9prpt5trx