[IMP] edi: work-in-progress: remove permanent storage of edi.document
The EDI documents will now be generated on demand
and available from the Portal view of each document.
Instead of getting a link to a statically generated
EDI document, customers will receive a link to
the portal access to the document. They will be able
to signup on the portal the first time as well,
provided they are using the secure token that was
sent to them (i.e. the right link).

The link to pay online will be available in the
portal as well.

Still much to do, this is a small first step,
with edi.document renamed to edi.edi for
consistency, as it will not persist any
edi.document anymore.

bzr revid: odo@openerp.com-20121011152008-bht7ub6woaex0a7u