[IMP] readonly sortable o2m
The making of the column invisible would happen after the header was
rendered (?), and the header would not be re-rendered afterwards. So
there would be a missing chunk of table. Make the cells invisible
instead. It's all going to be changed looks-wise anyway so meh.

bzr revid: xmo@openerp.com-20121022103935-78bm40gfotpt1ds6