[FIX] res.lang: ar_AR does not exist (Arabic in Argentina?!) -> replaced with ar_SY
ar_SY is the default locale alias for the `ar`
code in babel, so we can use that until a
translation team asks for a regional
variation of Arabic.

This should also fix an issue in the web
client, where ar_AR was not working, because
silently replaced with ar_SY.

This will only work for new databases,
old ones should be manually fixed.

bzr revid: odo@openerp.com-20121024161527-vec2r46a4o5lk47x