[FIX] modules: avoid to have the docutils' warnings rendered inline.
The html rendering of the __openerp__.py description is done using docutils.
This patch removes the warnings that docutils inserts in the resulting document
and log them. Docutils still generates spurious output on stderr though.
(An alternative is to use the `report_level` option but in that case it seems
we can never witness the warnings being created.)

bzr revid: vmt@openerp.com-20121025130111-7l157hexh6ew9w34