[FIX] res.partner: propagate `is_supplier` property when creating parent or child partner for a supplier
This is not necessary for the `is_customer` flag, as it is
always True by default (customers are created more frequently).
Passing the field value via an invisible field in the mini
contact form is necessary because context `default_*` values are
automatically discarded by the ORM when creating o2m
or m2m records (as they are supposed to apply to a different model).

bzr revid: odo@openerp.com-20130416091027-y7iwvpjhowg53lry