[IMP] simplify ir.ui.view.get_root_ancestor by using browse records
Also only case which should result in the id not existing is the
initial record, if a view_id is explicitly provided. So the loop can
avoid it's, it's traversing through an m2o so if the m2o value is not
null the next record in the chain should always exist.

bzr revid: xmo@openerp.com-20130423135739-jve1fe2it8q4gkwh