[FIX] rml2pdf: prevent report expression evaluation from crashing report completely
This is a revert of revision 4838 revid:vmt@openerp.com-20130327160420-pxw83283xcro1h52
which was based on the idea that we ought to fix broken reports rather than hide
their problems.
This is correct but is blocking users even for tiny details in the report, so
it is better to attempt generating the output anyway, and log an obvious error
so we have all the data needed to fix the report.

bzr revid: odo@openerp.com-20130627114916-d5p1qjye9tsgammj