[FIX] reinstate basic grouping test, replace strictEqual call on widgets by ok(===)
There seems to be a problem with QUnit's serialization routines when
applied to complex widgets (the view here) while running inside
PhantomJS. It does not seem to be cycles in datastructures (that is
handled) but looks related to complex objects linked from multiple
"children" (skipping any object already serialized makes the issue
disappear).

The bug leads to unbounded memory growth, and on runbot to the process
being summarily terminated (likely by the OOM killer) not even
resulting in an OOM error (which is displayed locally). The issue can
not be observed within browsers. This may be a difference in GC
strategy.

The serialization is applied by the driver on high-level comparison
methods (e.g. strictEqual) so the caller can display the structures if
desired.

Replace the strictEqual(a, b) call by ok(a === b) to avoid any risk of
such serialization.

bzr revid: xmo@openerp.com-20130801091326-u2q9e163zls4k8ad