[FIX] wrap the wrapper in order to allow RTE on all of the page (except the toolbar)
CKEditor does not work directly on body, and it's a bad idea anyway
(editing the edition toolbar won't end well). 

A big wrapper around all of the page's content can be used as RTE hook
to give access to the body as well as the header and footers.

bzr revid: xmo@openerp.com-20130918145028-6ppi9sro2un0quh5