[REV] stages: removed bar_fold and bar_color fields.
After some experiments, having colors does not seem to give an
usable result on statusbar widget. Indeed we need to have the
current stage highlited, and maybe something to tell that some
stages have a special meaning and/or are the end of the pipe.
The closed field already gives that meaning.

Moreover the issue with the statusbar widget is that it indicates
state and action. Having both dimensions with colors / icons
is quite difficult to understand. We therefore stand on a simpler
version of the widget, only with stages and the current stage
highlighted.

bzr revid: tde@openerp.com-20131021104912-8ybhw0svdoghheh3