[FIX] sale_stock,stock: change model to stock.picking.out of deliveries to invoice view in sales
Deliveries to invoice in sales menu should display delivery order only (no incoming shipment). This was already the case thanks to the domain [('type','=','out')], but since the refactor of the module stock, and the division of stock.picking to stock.picking.in and stock.picking.out, the model of this view should be stock.picking.out instead of stock.picking (for instance, to get the actions binding (ir.values) of stock.picking.out model).
+ typo fix in action binding

bzr revid: dle@openerp.com-20131206111336-dg01y92jvjnxy5oi