[REVERT] rev 9007 dle@openerp.com-20131122150606-ruqe2l0e7gn84upf
default_model and default_res_id were not present when doing this fix, because in the web client, we prevented the propagation of the context keys beginning with "_default" in some cases.
The fact is that, in this specific case, the keys "default_" should not have been removed. This has been fix in web client in 
rev 3892 revid:chs@openerp.com-20131211163609-i3s2mlncf5n91uda

Besides, active_model cannot be used here, as active_model is not 'sale.order' but the model of the wizard.

bzr revid: dle@openerp.com-20131211171429-6bxuh7o4ueiy9dd1