[FIX] hr_timesheet_sheet: save correctly if focus is in a timesheet input
on_change_unit_amount was called manually in the timesheet widget. Normally, save waits for all on_change to be applied, but in this case, as the on_change its called manually in the js, it was not the case.
The fix is to put all deferreds for which we must wait for in an array, and to wait all deferreds to be done before saving.

bzr revid: dle@openerp.com-20140115165114-nqa2ynszgocrnt52