[IMP] website_mail: avoid marking old messages as published
There is semantically no reason to publish old
messages, and doing so can be expensive for large
databases. Defaulting new messages to published
makes sense depending on the model on which they
are attached, so it might be better to improve
this as an opt-in handled in create() or
message_post() or via default_* context keys.

bzr revid: odo@openerp.com-20140217142006-ik18uqy9ewhhhb42