[IMP] improves searchview appendTo
after some thought, the extra method insert was not really necessary.
It was at first a way to force the user to give two nodes to determine
where the searchview and the drawer were to be inserted.  However,
the second node can be omitted, a perfectly reasonable default is
to append the drawer just after the searchview.

Then, the role of insert is exactly the same as appendTo, so it makes
sense to override appendTo and remove insert.  Simpler, and the job is done.