[IMP] change scrollable area to main view
Now, only the main view is supposed to scroll (the part below the
header with the view selector).  The left menu, top bar and header
are supposed to stay fixed.  This is done by carefully using
position:relative and absolute.  However, this is likely to
introduce some (many?) issues, perhaps with tooltips position and
such.  The bug hunt is on!