[FIX] web: on launching ir_actions_act_window, reset the active_* params if the target is current
This commit is related to 8d496399336b4a97200f6f7b5907ed1c7b99970c and b88755c4317ac61fc8d0533162d4ed19082f3fab
When hitting buttons of type object, the active_model and active_id(s) were kept, and, therefore, when calling a feature using the active* args, this feature used the active* args from the previous action.
Nevertheless, concerning wizards, the active* args should be indeed the active* args of the previous action, as wizards expects to have the active* args from the previous action. Thus, we reset these active* args only when this is not a wizard (target === 'current')
For example, from a sales order, hit the 'view invoice' button, and on the invoice, hit the 'send by email' button: The active_id in the send by email wizard were the id of the sale order, not of the invoice