[FIX] Partner VAT Intra XML delaration: Vat number format correction
The Partner VAT Intra XML delaration needs to have the VAT number
in the format without the country code prefix.

The current behaviour was to include the country code prefix in the
report which causes an SQLException error on uploading the XML file
to the Federal goverment's website.  While the error is non-descriptive
fixing the VAT number format fixes the error.